Reverted back to making inner loop always over i-index.
Tests showed that on cpus, using 1D pencils in j- and k-sweeps
slowed the code down by 50%.  Fund that by switching to level=0
scratch memory, and reducing the number of registers used in kernel
the orignial inner-i-index loops design was as fast as faster on GPUs.
Will try to optimize this approach now.