Changes to lots of files to add scalars to Hydro class.
Realized it was better to add scalars directly to Hydro rather than adding
as seperate class as this reduces a lot of duplication of code, and most
importantly it allows seperate scalars for hydro and MHD in two fluid simulations,
e.g. different species of neutrals and ions.

Code compiles and runs LW implosion test successfully.