Switched loops in DivFlux function to always run over x1.
This is the way the loops are organized in Athena++.  I had
implemented it earlier, but it performed poorly with small MeshBlocks.
With new update over all MBs, it does much better.  Will now try
comparing performance on GPUs and CPUs for the two approaches.
Downside of this appraoch is that three different reconstruction
functions are needed (one for each direction).