Cargrid output (#627)
* Fix interpolation on Chebyshev grid

Fix bug when interpolating on Chebyshev grid in which the indices were
computed with the Chebyshev grids, but the interpolation weights still
used the Cartesian grid

* Added new CartesianGridOuput output class

`CartesianGridOutput` can output any variable in a Cartesian box (within
the simulation domain) at a prescribed position and extent.

To use it, add a block to the output file like this:

```
        <output3>
        file_type  = cart
        variable   = hydro_w
        dt         = 0.1
        center_x   = 0.0
        center_y   = 0.0
        center_z   = 0.0
        extent_x   = 0.1
        extent_y   = 0.1
        extent_z   = 0.2
        chebyshev  = 1      # 0:uniform grid; 1:Chebyshev-Lobatto grid
```

Note that the extent of the grid is actually half of the extent
variable. This is because we are using the same convention as in
`CartesianGrid`.

This commit also adds `vis/python/cartgrid.py` to read the data produced
by `CartesianGridOutput`

* Add support for derived variables in CartesianGridOutput

* Fix CartesianGridOutput with AMR

Reset CartesianGrid interpolation indices and weights before
interpolation if AMR is enabled

* Fix bug introduced in previous commit

* Fix typo introduced in merge