Restart debugging and fix.
I noticed that the esrcs masking was not working correctly when the code
restarted from a .rst dump. This was because the user parameters was
being read in after the restart flag in the user pgen script. Changing
the order makes the results consistent whether run from restart or from
initial run.